move repository to under /yatai
